trying to resolve conflict in cmake install.
The default cmake install has a cmake.m4, which apparently
is in the cmake-3.7.1 package. When installing the new cmake-3.17.2
package, this file comes as part of cmake-data-3.17.2 which is
installed as a dependency to cmake.
Apt-get install complains about this file even though it will be
the same state afterwards.

If this fails, we'll also have to apt-get remove cmake before.

Signed-off-by: Henner Zeller <h.zeller@acm.org>