Continue regression test even if uhdm-dump fails.
I've seen udhm-dump fail for me on a couple of tests, but the single-threaded
regression run then stops immediately when the first crash happens.

Catch the failure so that we can continue safely.

(Don't know yet what is wrong with the dump in the first place, that
will be another patch :) ).

Signed-off-by: Henner Zeller <h.zeller@acm.org>