Be more explicit in generated artifacts by add_custom_command().
This now explicitly lists the artifacts generated by the parser generator
or the source-code generating Tcl scripts. This allows cmake to
provide a much more finer grained dependency generation as well as
making sure that dependencies of deleted generated artifacts are
properly generated (e.g. fixes #99).

Also this means that it is clearly documented in the build system
already for a contributor.

The generate_parser_listener.tcl script has been modified to
take an optional output directory; this is in prepration to
eventually output the generated artifacts into the build directory
and not mess with the original source. As of now, this is not
used yet, so right now, this behaves exactly as before.

Signed-off-by: Henner Zeller <h.zeller@acm.org>