Issue #239 (Part 3) Add Windows platform support
Highlights -
Raise minimum required CMake version to 3.15. This is required
for a successful build on Windows. This version introduces a
way to enforce which runtime library (/MT(d) vs. /MD(d))  gets
used across all subprojects. Without this feature each subproject
has to have a way to either set this up from parent project or
has to be modified to be built for our requirement. At of this
writing neither flatbuffer nor googletest have a configurable
option.