Start using std::string_view in StringUtils.
std::string_view has a multitude of advantages; most specifically, it
allows to write code that prevents copies of strings.
Gently transform a first set of functions for that.

There is more opportunity, but I didn't change the functions itself
too much, as there are no unit tests, so this would be dangerous.

Using std::string_view consequently everywhere it is possible (with
proper string ownership known), it is possible to hand around
string slices without ever to copy strings.

string_view appeared in C++17, since we're using this now as minimum
requirement, this was possible (otherwise we could've used
absl::string_view).

Signed-off-by: Henner Zeller <h.zeller@acm.org>