Preparation to fixing some memory leaks.
The unit tests are leaking a lot of memory, so in an attempt
to drill down on these, these are a few initial changes.

This is mostly not doing any functional change yet, but merely
adding some clarity in some places, making things const where
possible, adding TODOs to look into. Sent as pull request
now so that it is small for quick review.

Signed-off-by: Henner Zeller <h.zeller@acm.org>