Fix delete of design by non-owner.
The Design* owned by the Compiler was deleted in main while it
needs to be deleted by Compiler.
Removed the delete there and enable the deletes in the Compiler
destructor.

While at it, add a todo to not hand out writable Design*, but
that will be a larger, long-term effort.

With this change, two more of the unit tests pass without leaks.

Signed-off-by: Henner Zeller <h.zeller@acm.org>