Add test for StringUtils::removeComments()
It was not tested in the unit-test.
Also, add comment in the StringUtils.h what this function
is doing (as its naming implies removing all comments, but
in actuality it removes certain kind of end-of-line comments).

Document in test the things it does not do yet: ignore comment
characters in string constants.

Signed-off-by: Henner Zeller <h.zeller@acm.org>