Make whitespace-matching regular expression more robust.
The regular expression matching whitespace in PreprocessFile is
matching multiple whitespace characters, including zero. This is
ambiguous depending on the greediness of the implementation, as
it can match any empty sub-string. Change the expression to
mean 'at least one'.

While at it, pre-compile the regex instead of doing it every time
in the loop.

Signed-off-by: Henner Zeller <h.zeller@acm.org>