Issue #1718, #1727 Surelog parsing is pretty slow
Resolution
Replacing use of standard RTTI i.e. use of dynamic_cast
with a custom implementation that avoid having to enable
RTTI across the entire application. This changes replaces
it for antlr only and future change would do the same
for whole of Surelog.

NOTE: The compile provided RTTI is not yet disabled.