Issue #1718: Fallback to pristine antlr
Introduced RTTI solution seems to be slower than the standard
language implementation using dynamic_cast. There is no reason
for this to be the case and needs investigation. For the time
being, as a stop gap solution, falling back to pristine Antlr
to not have all developers pay the penalty.