Memory Optimization Pass
Use std::string_view in place of std::string in SymbolFactory.
Updated all code generation scripts to use std::string_view.

Updated build script to build googletest with C++17 as well so
we are using the same compiler settings across the entire project.
This is also required to bring in support for std::string_view.