Add a clang-tidy run to the CI
Using a baseline of useful clang-tidy warnings, but with
some disabled that would be too noisy right now.

For now, this is just generating an FYI output as
a starting point to get these adressed. Once the baseline
is clean, we can change this to result in an CI error if
new warnings show up in a pull request.

Signed-off-by: Henner Zeller <h.zeller@acm.org>