Speeds up generate statements.
This PR addresses the exponential look ahead required to parse generate
statements in modules.

I addressed two things. The first is I updated a non-compliant grammar
rule. Then added semantic predicates to the generate_item rule that
slowed down parsing signficantly.

The SV Spec states

```
Within an interface_declaration, it shall only be legal for a generate_item to be an interface_or_generate_item.
Within a module_declaration, except when also within an interface_declaration, it shall only be legal for a
generate_item to be a module_or_generate_item. Within a checker_declaration, it shall only be legal for a
generate_item to be a checker_or_generate_item.
```

These semantic predicates enforce that footnote in the spec. In one of
Google's internal projects we found a file that was would run for 20+
hours on current head and never complete. This PR parses that file in
just 2 minutes which is a signficant improvement.