Enable clang-tidy for selected findings; fix these findings.
To have the clang-tidy run have any impact, we need to fix the
findings to get the code to improve. Since there are many hundred
findings currently, let's do this one-by-one.

In this round: fix override/final.
Then enable clang-tidy run for just this particular warning and
make this fail in the CI if new findings happen.

That way, we can over time fix one thing at the time and then
enable the corresponding check to keep in check in the future.

To run clang-tidy with all findings, simply run
  .github/bin/run-clang-tidy.sh

.. and to only run with the limnited set of findings run
  .github/bin/run-clang-tidy.sh limited
(this is what the CI is running)

Signed-off-by: Henner Zeller <h.zeller@acm.org>