Introducing RoundTrip - application to validate location info
RoundTrip is an application to use the location information from
generated model to regenerate the sources - making it a round trip. The
geneated sources are compared against the original to generate statistics.
The application can be run in two independent modes - surelog & uhdm. The
former, Surelog mode includes parsing and the later from binary. RoundTrip
is also part of regression and statistics are reported with every run.
However, the differences (or rather change in differences) wouldn't
report a failure.

The application can also serve as a initial model for implementing a
linter.

P.S. This is still a work in progress. A number of open issues where
the required information is missing or in other cases wrong.