Don't copy all exisiting entries on every copy constructed SymbolTable.
The SymbolTable's copy constructor is used in the code as snapshotting
the current state, but that is resulting in a huge churn in memory as
a lot has to be copied on each such occasions.

This change makes it cheap as snapshotting starts with empty local
containers and then delegates lookups to parents as needed.
This requires an indirection to the parent if there is one, but this
saves a large amount of memory churn and runtime.

This change saved many gigabytes of memory use and a lot of runtime
while processing a larger project. So very worth-while.

Longer term, all the places that use the SymbolTable's copy constructor
to create a snapshot should be investigated if they can be
solved differently.

Siblings of the same parent are thread safe and can operate concurrently.
This is _not_ thread safe if the parent is modified while the client
does a look-up.

There are simple ways around it, but currently, threaded compiling is
[not done](https://github.com/chipsalliance/Surelog/blob/dc9970a0827a84ce5d7dca60f2264d32503a37f3/src/DesignCompile/CompileDesign.cpp#L282-L288),
so this is not an issue right now. And by the time we can do
multi-threaded compiling again, I hope to have resolved the
SymbolTable copy-need already. So for sake of simplicity, no effort
is put into thread-safety right now.

TESTED
Added comprehensive unit test to make sure that all properties of
the previous symbol table still work and behave exactly the same.
The regreession test passes with no diffs. But faster and using
less memory :)

Signed-off-by: Henner Zeller <h.zeller@acm.org>