Introduce StrCat() and StrAppend()
These functions allow to concatenate arbitrary set of arguments
to a string. These functions can now be used in contexts where
strings are concatenated, but some arguments (such as integers)
require to call std::to_string - these are now much easier to
be formulated.

Main use: whenever there are strings concatenated, prefering
this will allow even faster string concatenation in the future
(the implementation is quite simple right now, but it is feasible
to replace this with something that pre-calculates the buffer to
be allocated).
Also, this is compatible with std::string_view, which otherwise
are a pain to += concatenate.

Added StrCat(), StrAppend() and unittest.
As a first user and example, convert the string-concatenation
in VObject::print()

Signed-off-by: Henner Zeller <h.zeller@acm.org>