Only store symbols that are referenced in the cache.
Only store relevant symbols for a particular cache emit: instead
of including the entire symbol table in the cache output, build
up the canonical cache symbol table with IDs we see and only emit
these.

In the course of that, discovered and fixed a bug: the VObject fileId
was not re-mapped to the new symbol table ID, but was using the
local symbol table ID which is of course not valid after restore.

Signed-off-by: Henner Zeller <h.zeller@acm.org>