Issue #2962: Disabling Rsd
Early investigation is pointing to Rsd test being the culprit for
stalling the regression on CI. The root cause of why isn't identified
yet but disabling the test to see if this fixes the next few builds.
This is by all means is a temporary solution only to confirm whether or
not this is the root cause. There are problems identified in the Rsd.log
which might be more clue to fixing the test itself rather than disabling
it.