Shorterm fix for WinClang stack overflow issue
The stack overflow was caused by the check loop functionality for
evaluating/reducing expressions. This needs a better logic than merely
depending on a stack frame counter.

Generally, for windows, the stack size has to be set for the application
but for some unknown/unidentified reason the stack usage for Clang seems
unnecessarily high. The set limit for the recursion was 1000 but even
with 64MB stack size the application kept running into the same
problem. Probably some other undelying issue (potential large objects
crated on stack) but a problem for some other day.
Implementing a short term solution to get the builds working again.

In the process, also found that the same problem happens with MSVC cl
compiler as well. In debug mode, the application blows up the stack.

Short term solution - For Windows only, reduce the limit from 1000 down to
100 and set the stack size to 32MB for clang builds.

A follow up problem - The value post reduction failure was invariably
dependent on the maximum loop counter. Even though the builds and tests
worked independently for the platform, the results were different across
platforms because of different max values for the loop counter. To
resolve this difference in resultant value post failure, use the input
unmodified.