Give user a choice to use flatbuffers from local system.
Providing SURELOG_USE_HOST_FLATBUFFERS to cmake.
Default OFF. If ON, flatbuffers from third_party/ is
skipped but rather attempted to be searched on the
host system using pkg-config.

This is the first step, similar options need to be
provided for capnp, antlr and uhdm.

This will make it much easier for packagers (e.g. Debian)
to lego-assemble the relevant components fitting into the
system.

Signed-off-by: Henner Zeller <h.zeller@acm.org>