Print formatted object stats during regression runs
Regression runs used merely the count i.e. number of lines of UHDM dump
output to identify differences. This has proved very unreliable with
some of the recent changes where the number of output lines are exactly
the same but the type of UHDM model has changed (for instance, from
int_typespec to logic_typespec or vice versa).

To make the reporting a little more reliable, force dumping of object
stats and compare them.

Until logs are updated, missing stats dump in golden will not report an
error (to prevent immediate failures because of this change).