Issue #3187: Add support for using zip file
The files downloaded from CI builds are zip files (only option from
Github). There was a manual step to unzip this to pull out the tar.gz
which was then used by the extract mode.

Added support to use the zip file directly without the manual step.
Note that both zipfile and tarfile are default modules with python
installation so there is no additional dependency required.