IO Abstraction: Introducing PathId & FileSystem (Phase 1)
NOTE: This is only the first phase of the large goal of IO abstraction.

Big picture goal is to abstract IO operations out of Surelog and into a
class implementation that can be overriden by client application. This
facilitates implementation of other more extensive features (like
support for UNC paths, portale cache, multi-source root directories,
cache archives, etc). Adding those directly to Surelog is not ideal
since in many cases this features pull in other thrid-party dependecnies
that might not have the same licensing terms as Surelog (and so leave it
to the client/application to decide).

Introducing two new classes - PathId & FileSystem

PathId - New class parallel to SymbolId that represent path-like objects
for IO operations. Each instance can represent a directory or a file but
the internal representation doesn't necessarily have to be a _path_. For
instance, in case of compressed archive of sources, it can be a string
representation of the archive + the offset. The only enforcement is that
the representation needs to be *printable* i.e. should be convertible to
a string (since thee representation is stored in SymbolTable). For the
advanced users. this limitation can very easily be overcome and is left
as an exercise.

FileSystem - This reroutes access to native file system or in special
advanced cases can restrict access to the file system. All operations of
PathId are meant to be routed thru' this class. This commit contains
only the bare-bones of the implementation (to keep the rest of the
change review-able). Follow up commit will add the real meat to this
implementation.

Most of this change is about switching from SymbolId to use of PathId
where applicable. Some specific ones that may be of interest to the
reviewer.

* Restrict use of std::filesystem to minimal scope (functions where
  possible). Do NOT pass std::filesystem::path across function boundaries.
  Use PathId instead.
* Do NOT cache unpinned strings, especially the ones that represent
  paths. Add them to the SymbolTable and resolve it on restore.
* PreprocessFile::m_fileId used to represent both a file path (when
  processing a file) or a macro name (when processing a macro). This was a
  problem since now the path needs to be represented using PathId but the
  macro name is represented using SymbolId. So, added two independent
  variables PreprocessFile::m_macroId & PreprocessFile::m_fileId.