Add more meaningful timeouts for jobs on CI
When GitHub force kills the workflow after 360 minutes, it's a hard
reset i.e. no steps are allowed to run. This hard reset prevents the
artifacts upload as well making it impossible to diagose the underlying
issue. Adding a more reasonable timeout for long running steps in the
workflow so that it leaves enough time for the workflow to do the
trailing steps.

For now, adding a strict timeout of 180 for regression test runs. This
is the longest running step and most likely to hang/spin infinitely
because some or other test run got stuck.