Issue #3239: Remove errorneous error reporting
Apparently, it's possible that the start and end file of a VObject token
are different. One such case is when a include file is in the middle of
a module declaration.

Known Issue: VObject still needs to track the end file but there's only
one PathId member which is tracking the start file. Something to follow
up on.