Workaround for windows console buffer limit
Console on Windows has a maximum buffer limit of 9999 which isn't enough
when working with some of the large tests that generate significant
amount of data to the output (stdout and stderr) streams. Add a windows
specific workaround to redirect the output to a file when running under a
debugger.