Handle carriage return and newlines across platform boundaries
When a repository is sync-ed on Windows, the default line-ending used by
Git is \r\n. But when building under WSL, the line-ending the code
expects is \n. This difference causes the generated chunk files to be
different across the platform making identifying differences harder.

Small edit in code to remove any carriage returns at the end of the line
for content used to generate chunk files.