Fixing a race condition during regression
BuildOVMPkg & BuildUVMPkg tests had the -createcache option which would
mean the Surelog binary writes its cache output to the precompiled
folder. This creates a race condition because as some other test could
be reading the same file. Removing the flag and letting the tests write
to its own designated folder.