IO Abstraction (Phase 7): Make cache portable
Make caches usable if either the sources or the cache itself is relocated.
This enables caches generated on one machine to be usable on other and also
across platforms (i.e. cache generated on Linux to be used on Windows). This
is, however, achieved only via manual input thru' command line.

* Add new command line option "-remap" to accept path remapping
  arguments
* Added a test in FileSystem to verify the feature
* Updated both PPCache & ParseCache implementation to remap loaded paths
  before registering them with the SymbolTable or using it otherwise.