Don't use -DDEBUG to change layout of PathId.
Since -DDEBUG is commonly used, it is dangerous to change the layout of an object just because of that - project using the Surelog library will now accidentally change the size that is not compatible anymore with the binary library.
So, only do this if the very specific debugging option for this case is set.

Related: https://github.com/chipsalliance/UHDM/issues/872