USE_HOST_UHDM: Get udhm-dump from path instead of third_party/
If the user chose SURELOG_USE_HOST_UHDM, there will be not uhdm-dump
binary build in third_party/ but instead, it will be available in
the users' environment.

So add this fallback: if the uhdm-dump binary is not found in the
third_party/ location, take it from the PATH.

The change in the tcl file looks pretty big, but it is only three
lines; the rest is fixing trailing whitespace.