Avoid cycle in VpiListener logic and during binding
listenDesign_ clears the visited container between iterating allXXX and
topXXX. That removes the top level design as well (which was added by
listenDesign) causing infinite loops.

This is found in cases where "$root" is used and the corresponding
ref_obj::actual_group points to a design.

A ref_obj shouldn't point to another ref_obj as its actual_group.