Don't refer to project headers in *.cpp files as system headers.
Just textual update of includes in *.cpp files, no code changes.

This implements style-guide rule
https://google.github.io/styleguide/cppguide.html#Names_and_Order_of_Includes

We do use angle bracktes for inlucdes within the headers that are
also system-installed with angle brackets so that they can be used
in such context. This is the reason why angle brackets made it into
the project.

However, within our own _implementation_ we should refer to the local
project headers as such, with #include "quoted.h" instead of
angled.

Within each compilation unit, the corresponding declaration header
should be the very first header (that way we can detect quickly if
they are self-contained).

Both of these issues are addressed here.

The includes in _header_ files are not addressed at this point,
they are still included as system headers to accomodate the original
use-case of installing them system-wide.