New API function to compare parser outputs from two compilers
The API compares the AST trees generated by the invocation of two
compilations calls, possibly on the same/similar inputs. Spaces (both
white space and new lines) are ignored in comparison. Note that is a
topological comparison, so based solely on the "type" of the node and
nothing else.

Intention is to use this API to validate the code generated by another
tool (like a formatter) and comparing the AST of original to the new
source.