fix: don't try to attribute names to ordered ports
The code was attributing name to ordered ports based on the Primary
literal passed to the port of the module instantiation (if any), which
is not the name of the port. I believe the port name is later corrected
when the module is resolve to its declaration. But if the module is not
resolve (is a unkown module) the generated UHDM would have the wrong
name for the ports. It is better to keep the port unnamed.