Fix handling of line comments in multi-line macros
Grammar for multiline macros was over aggressive causing out of order
enter/leave calls for the listener. The problem stemmed from the
ambiguity between escaped carriage return and line comment both
expecting a new line at end (except one was escaped and other wasn't).
Lexer, however, didn't have the knowledge of this difference and so the
tokens generated caused the parser to over consume for the multi-line
macro.

To resolve, introducing a new token to represent a escaped line comment
and use it in place of regular line comment within multi-line macros.
The only difference between the two is the former being terminated by an
escaped newline rather than merely a newline.